+++ src/apiCalls/economics/index.ts
@@ -1,0 +1,1 @@
+export * from './getEconomicsInfo';